refactor: 유저 정보를 가져오는 CustomUserDetails 클래스 리팩토링
