+++ src/main/java/balancetalk/global/config/CustomUserDetails.java
@@ -1,20 +1,63 @@
 package balancetalk.global.config;
 
+import balancetalk.member.domain.Member;
+import java.util.ArrayList;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-
 import java.util.Collection;
+import org.springframework.security.core.userdetails.UserDetails;
 
-public class CustomUserDetails extends User {
+public class CustomUserDetails implements UserDetails {
 
-    private final Long memberId;
+    private final Member member;
 
-    public CustomUserDetails(String username, String password, Collection<? extends GrantedAuthority> authorities, Long memberId) {
-        super(username, password, authorities);
-        this.memberId = memberId;
+    public CustomUserDetails(Member member) {
+        this.member = member;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        Collection<GrantedAuthority> collection = new ArrayList<>();
+        collection.add(new GrantedAuthority() {
+            @Override
+            public String getAuthority() {
+                String role = String.valueOf(member.getRole());
+                return role;
+            }
+        });
+        return collection;
+    }
+
+    @Override
+    public String getPassword() {
+        return member.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return member.getEmail();
     }
 
     public Long getMemberId() {
-        return memberId;
+        return member.getId();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
     }
 }
