+++ src/main/java/balancetalk/global/config/MyUserDetailService.java
@@ -5,13 +5,10 @@ import balancetalk.global.exception.ErrorCode;
 import balancetalk.member.domain.Member;
 import balancetalk.member.domain.MemberRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import java.util.Collections;
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -23,15 +20,6 @@ public class MyUserDetailService implements UserDetailsService {
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         Member member = memberRepository.findByEmail(username)
                 .orElseThrow(() -> new BalanceTalkException(ErrorCode.NOT_FOUND_MEMBER));
-
-        List<GrantedAuthority> authorities = Collections.emptyList();
-
-        CustomUserDetails userDetails = new CustomUserDetails(
-                member.getEmail(),
-                member.getPassword(),
-                authorities,
-                member.getId() // memberId를 추가
-        );
-        return userDetails;
+        return new CustomUserDetails(member);
     }
 }
